pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llmwiki"
version = "0.1.0"
authors = [
  { name="Lucius", email="your.email@example.com" }
]
description = "A lightweight CLI tool for building and maintaining LLM-powered knowledge bases"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "click>=8.0",
  "openai>=1.0",
  "anthropic>=0.20",
  "pydantic>=2.0",
  "python-dotenv>=1.0",
  "PyPDF2>=3.0",
  "python-docx>=1.1",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "openai-whisper>=20231117",
  "whoosh>=2.7.4",
  "GitPython>=3.1",
  "python-slugify>=8.0",
  "markdown>=3.5",
  "python-dateutil>=2.8"
]

[project.scripts]
llmwiki = "llmwiki.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["llmwiki*"]
